This is C++ code, not C code
Unlike printf, cout << endl can allow for both LF and CRLF with absolutely no change to the code. In the case of printf, if you want to compile this command for Windows, you'd have to create a Windows version with "\r\n" and a Linux version with just "\n" in lines 8 and 10.